requirements.txt
langchain>=0.0.200
sentence-transformers>=2.2.2
pinecone>=2.1.1
pypdf>=3.7.0
python-dotenv>=1.0.0

# Backend - FastAPI & Perplexity
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.31.0

# Production server
gunicorn>=21.0.0